requirements.txt
PyYAML>=6.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
torch>=2.0.0
networkx>=3.1
requests>=2.31.0
tqdm>=4.65.0
langgraph>=0.0.50
langchain-core>=0.1.0
openai>=1.0.0
ollama>=0.1.0
